python/requirements.txt
psycopg2-binary
requests
qdrant_client